Install Devise, and Create the Admin Resource
This commit includes installing the `devise` and `devise_invitable` gems,
and creating the Admin resource. I also generated the Devise-default view
files, just in case we'd like to edit them in the future.

Note that there is a database migration to run with this commit, to create
the Admin resource in the database.