Stripe Integration Deploy 1 - Plan
Integration with Stripe Plans, including revamping the Plan resource to
interact with Stripe.

This also includes some migration scripts to get the local DB set up
for mirroring Stripe Plans.

The DB migration adds `stripe_id`, `category`, and `category_order` to
the `Plan` model, and also adds `stripe_id` to the `Member` model to
prepare for the next deployment.