Convert Trial Period Days to Calendar
This allows admins to select the actual end date of the trial period,
instead of entering in the number of days in the trial period.

Note that the trial end date is not based on the date of signup, like
instantly-paid subscriptions. Thanks to timezone-manipulation issues and
Heroku preferring UTC-based timestamps, I have to set the trial end
timestamp to 12pm UTC in order to get the date to actually set correctly
in Stripe.