Set Start Date based on Trial End, Hide Start and End Dates for Stripe
If a trial end date is set, we'll set the start date to that date. We'll
also show the membership status as "trialling" instead of "live".

Also, if the membership is being paid through Stripe, we'll hide the
"Start Date" and "End Date" fields, since they're redundant at this
point.