Dynamically add rows to form
Instead of hard-coding six (or any number of) rows, I've left a single
row on the form, and added an "Add Row" button to dynamically add
additional rows.

I also had to change the `change` handlers to be bound to `document`, so
they'll pick up events on the dynamically added rows. The form
serialization just needed to be bound to the form ID instead of the
`form` tag itself for the same reason.