Add Proration to `CustomSub#edit` and `#update`
I've enabled the option to prorate the subscription changes. Prorating
adds the difference of the changes between now and the end of the
billing period to the next invoice; not prorating just means that the
changes "take effect" with the next invoice.

It should be noted that, for simplicity sake, the "proration date" is
12am PT of the current day. This provides consistency between previewing
the updates, and actually committing them, as well as committing
multiple updates in a single day.