Quick fixes for Stripe Sidekiq jobs
- `update_stripe_data_worker` - apparently the 
`stripe_next_inv.period_start` field isn't properly being fetched and 
returned by Stripe's backend (or so I gather from a chat with their 
support team), so we'll grab it from `date` instead. This seems to be 
the field that the Stripe Dashboard is pulling from for its Upcoming 
Invoice display, so this seems pretty safe.
- `check_renewal_window_worker` - add a quick nilcheck, since it seems 
that this is failing because of it.